requirements.txt
fastapi
uvicorn
python-dotenv
pandas
numpy
nltk
sentence-transformers
scikit-learn
joblib
torch
PyGithub
tabulate
lightgbm
